refactor: split internal/commands into focused sub-packages
Decompose the monolithic internal/commands package into six focused
sub-packages (addon, cluster, ctxcmd, factory, nodegroup, workload),
each with its own command registration, actions, formatters, and tests.

Removes ~4800 lines of flat files in favor of co-located, package-scoped
code. Tests are updated to accurately reflect what each formatter
actually captures (pterm/color.Yellow write to the original stdout
handle, not the swapped pipe) with comments documenting the limitation.
