--- internal/commands/list_nodegroups.go
@@ -1,305 +1,0 @@
-package commands
-
-import (
-	"context"
-	"encoding/json"
-	"fmt"
-	"log/slog"
-	"os"
-	"sort"
-	"strings"
-	"time"
-
-	"github.com/dantech2000/refresh/internal/awsconfig"
-	"github.com/fatih/color"
-	"github.com/urfave/cli/v2"
-
-	"gopkg.in/yaml.v3"
-
-	awsinternal "github.com/dantech2000/refresh/internal/aws"
-	appconfig "github.com/dantech2000/refresh/internal/config"
-	"github.com/dantech2000/refresh/internal/services/cluster"
-	"github.com/dantech2000/refresh/internal/services/nodegroup"
-	"github.com/dantech2000/refresh/internal/ui"
-)
-
-// ListNodegroupsCommand creates the list-nodegroups command
-func ListNodegroupsCommand() *cli.Command {
-	return &cli.Command{
-		Name:      "list-nodegroups",
-		Usage:     "List nodegroups in a cluster with optional health/cost/utilization",
-		ArgsUsage: "[cluster]",
-		Flags: []cli.Flag{
-			&cli.DurationFlag{
-				Name:    "timeout",
-				Aliases: []string{"t"},
-				Usage:   "Operation timeout (e.g. 60s, 2m)",
-				Value:   appconfig.DefaultTimeout,
-				EnvVars: []string{"REFRESH_TIMEOUT"},
-			},
-			&cli.StringFlag{
-				Name:     "cluster",
-				Aliases:  []string{"c"},
-				Usage:    "EKS cluster name or pattern",
-				Required: false,
-			},
-			&cli.BoolFlag{
-				Name:    "show-health",
-				Aliases: []string{"H"},
-				Usage:   "Include health status for each nodegroup",
-				Value:   false,
-			},
-			&cli.BoolFlag{
-				Name:    "show-costs",
-				Aliases: []string{"C"},
-				Usage:   "Include cost analysis",
-				Value:   false,
-			},
-			&cli.BoolFlag{
-				Name:    "show-utilization",
-				Aliases: []string{"U"},
-				Usage:   "Include CPU utilization metrics",
-				Value:   false,
-			},
-			&cli.BoolFlag{
-				Name:    "show-instances",
-				Aliases: []string{"I"},
-				Usage:   "Include instance details",
-				Value:   false,
-			},
-			&cli.StringFlag{
-				Name:    "timeframe",
-				Aliases: []string{"T"},
-				Usage:   "Utilization window (1h,3h,24h)",
-				Value:   "24h",
-			},
-			&cli.StringFlag{
-				Name:    "format",
-				Aliases: []string{"o"},
-				Usage:   "Output format (table, json, yaml)",
-				Value:   "table",
-			},
-			&cli.StringFlag{
-				Name:  "sort",
-				Usage: "Sort by field: name,status,instance,nodes,cpu,cost",
-				Value: "name",
-			},
-			&cli.BoolFlag{
-				Name:  "desc",
-				Usage: "Sort descending",
-				Value: false,
-			},
-			&cli.StringSliceFlag{
-				Name:    "filter",
-				Aliases: []string{"f"},
-				Usage:   "Filter nodegroups (format: key=value, e.g., instanceType=m5.large)",
-			},
-		},
-		Action: func(c *cli.Context) error {
-			return runListNodegroups(c)
-		},
-	}
-}
-
-func runListNodegroups(c *cli.Context) error {
-	ctx, cancel := context.WithTimeout(context.Background(), c.Duration("timeout"))
-	defer cancel()
-
-	awsCfg, err := awsconfig.Load(ctx, c)
-	if err != nil {
-		color.Red("Failed to load AWS config: %v", err)
-		return err
-	}
-
-	if err := awsinternal.ValidateAWSCredentials(ctx, awsCfg); err != nil {
-		color.Red("%v", err)
-		ui.Outln()
-		awsinternal.PrintCredentialHelp()
-		return fmt.Errorf("AWS credential validation failed")
-	}
-
-	// Positional cluster name support; list clusters if omitted
-	requested := c.Args().First()
-	if requested == "" {
-		requested = c.String("cluster")
-	}
-	if strings.TrimSpace(requested) == "" {
-		ui.Outln("No cluster specified. Available clusters:")
-		ui.Outln()
-		start := time.Now()
-		svcList := newClusterService(awsCfg, false, nil)
-		summaries, err := svcList.List(ctx, cluster.ListOptions{})
-		if err != nil {
-			return err
-		}
-		_ = outputClustersTable(summaries, time.Since(start), false, false)
-		return nil
-	}
-	clusterName, err := awsinternal.ClusterName(ctx, awsCfg, requested)
-	if err != nil {
-		return err
-	}
-
-	// Use Error level to avoid warnings cluttering spinner output
-	logger := slog.New(slog.NewTextHandler(os.Stderr, &slog.HandlerOptions{Level: slog.LevelError}))
-	// Health checker optional for list (disabled by default)
-	svc := newNodegroupService(awsCfg, false, logger)
-
-	filters := make(map[string]string)
-	for _, f := range c.StringSlice("filter") {
-		parts := strings.SplitN(f, "=", 2)
-		if len(parts) == 2 {
-			filters[parts[0]] = parts[1]
-		}
-	}
-	opts := nodegroup.ListOptions{
-		ShowHealth:      c.Bool("show-health"),
-		ShowCosts:       c.Bool("show-costs"),
-		ShowUtilization: c.Bool("show-utilization"),
-		ShowInstances:   c.Bool("show-instances"),
-		Filters:         filters,
-		Timeframe:       c.String("timeframe"),
-	}
-
-	spinner := ui.NewFunSpinnerForCategory("nodegroup")
-	if err := spinner.Start(); err != nil {
-		return err
-	}
-	defer spinner.Stop()
-
-	start := time.Now()
-	items, err := svc.List(ctx, clusterName, opts)
-	spinner.Success("Nodegroup information gathered!")
-	if err != nil {
-		return err
-	}
-
-	// Apply sorting for table and when unspecified
-	if strings.ToLower(c.String("format")) == "table" || c.String("format") == "" {
-		items = sortNodegroupSummaries(items, c.String("sort"), c.Bool("desc"))
-	}
-
-	switch strings.ToLower(c.String("format")) {
-	case "json":
-		enc := json.NewEncoder(os.Stdout)
-		enc.SetIndent("", "  ")
-		return enc.Encode(map[string]any{"cluster": clusterName, "nodegroups": items, "count": len(items)})
-	case "yaml":
-		enc := yaml.NewEncoder(os.Stdout)
-		enc.SetIndent(2)
-		defer func() { _ = enc.Close() }()
-		return enc.Encode(map[string]any{"cluster": clusterName, "nodegroups": items, "count": len(items)})
-	default:
-		return outputNodegroupsTableWithWindow(clusterName, c.String("timeframe"), items, time.Since(start), opts)
-	}
-}
-
-func outputNodegroupsTableWithWindow(clusterName, timeframe string, items []nodegroup.NodegroupSummary, elapsed time.Duration, opts nodegroup.ListOptions) error {
-	if len(items) == 0 {
-		color.Yellow("No nodegroups found for cluster: %s", clusterName)
-		return nil
-	}
-	ui.Outf("Nodegroups for cluster: %s\n", clusterName)
-
-	// Show retrieval time with optional utilization window
-	if opts.ShowUtilization {
-		ui.Outf("Retrieved in %s (utilization window %s)\n", color.GreenString("%.1fs", elapsed.Seconds()), timeframe)
-	} else {
-		ui.Outf("Retrieved in %s\n", color.GreenString("%.1fs", elapsed.Seconds()))
-	}
-
-	// Show which optional data is being requested
-	if opts.ShowUtilization || opts.ShowCosts {
-		var extras []string
-		if opts.ShowUtilization {
-			extras = append(extras, "CPU metrics")
-		}
-		if opts.ShowCosts {
-			extras = append(extras, "cost estimates")
-		}
-		ui.Outf("Including: %s\n", strings.Join(extras, ", "))
-	}
-	ui.Outln()
-
-	// Define base columns
-	columns := []ui.Column{
-		{Title: "NAME", Min: 4, Max: 60, Align: ui.AlignLeft},
-		{Title: "STATUS", Min: 10, Max: 0, Align: ui.AlignLeft},
-		{Title: "INSTANCE", Min: 10, Max: 0, Align: ui.AlignLeft},
-		{Title: "AMI STATUS", Min: 9, Max: 0, Align: ui.AlignLeft},
-		{Title: "NODES", Min: 7, Max: 0, Align: ui.AlignRight},
-	}
-
-	// Add optional columns based on flags
-	if opts.ShowUtilization {
-		columns = append(columns, ui.Column{Title: "CPU%", Min: 5, Max: 0, Align: ui.AlignRight})
-	}
-	if opts.ShowCosts {
-		columns = append(columns, ui.Column{Title: "COST/MO", Min: 8, Max: 0, Align: ui.AlignRight})
-	}
-
-	table := ui.NewPTable(columns, ui.WithPTableHeaderColor(func(s string) string { return color.CyanString(s) }))
-
-	for _, ng := range items {
-		// Build row with base columns
-		row := []string{
-			ng.Name,
-			ng.Status,
-			ng.InstanceType,
-			ng.AMIStatus.String(),
-			fmt.Sprintf("%d/%d", ng.ReadyNodes, ng.DesiredSize),
-		}
-
-		// Add optional columns
-		if opts.ShowUtilization {
-			cpu := "-"
-			if ng.Metrics.CPU > 0 {
-				cpu = fmt.Sprintf("%.0f%%", ng.Metrics.CPU)
-			}
-			row = append(row, cpu)
-		}
-		if opts.ShowCosts {
-			cost := "-"
-			if ng.Cost.Monthly > 0 {
-				cost = fmt.Sprintf("$%.0f", ng.Cost.Monthly)
-			}
-			row = append(row, cost)
-		}
-
-		table.AddRow(row...)
-	}
-	table.Render()
-	return nil
-}
-
-// deadcode: retained for reference; use outputNodegroupsTableWithWindow instead
-// func outputNodegroupsTable(clusterName string, items []nodegroup.NodegroupSummary, elapsed time.Duration) error {
-//     // Default rendering with 24h if caller didn't provide
-//     return outputNodegroupsTableWithWindow(clusterName, "24h", items, elapsed)
-// }
-
-// sortNodegroupSummaries sorts nodegroup list output
-func sortNodegroupSummaries(items []nodegroup.NodegroupSummary, key string, desc bool) []nodegroup.NodegroupSummary {
-	less := func(i, j int) bool { return false }
-	switch strings.ToLower(key) {
-	case "status":
-		less = func(i, j int) bool { return items[i].Status < items[j].Status }
-	case "instance":
-		less = func(i, j int) bool { return items[i].InstanceType < items[j].InstanceType }
-	case "nodes":
-		less = func(i, j int) bool { return items[i].ReadyNodes < items[j].ReadyNodes }
-	case "cpu":
-		less = func(i, j int) bool { return items[i].Metrics.CPU < items[j].Metrics.CPU }
-	case "cost":
-		less = func(i, j int) bool { return items[i].Cost.Monthly < items[j].Cost.Monthly }
-	default: // name
-		less = func(i, j int) bool { return items[i].Name < items[j].Name }
-	}
-	sort.SliceStable(items, func(i, j int) bool {
-		if desc {
-			return !less(i, j)
-		}
-		return less(i, j)
-	})
-	return items
-}
